+++ invenio_rdm_records/requests/community_inclusion.py
@@ -9,8 +9,15 @@
 
 from invenio_drafts_resources.services.records.uow import ParentRecordCommitOp
 from invenio_i18n import lazy_gettext as _
+from invenio_notifications.services.uow import NotificationOp
 from invenio_records_resources.services.uow import RecordIndexOp
 from invenio_requests.customizations import RequestType, actions
+from invenio_requests.errors import CannotExecuteActionError
+
+from invenio_rdm_records.notifications.builders import (
+    CommunityInclusionAcceptNotificationBuilder,
+)
+from invenio_rdm_records.services.errors import InvalidAccessRestrictions
 
 from ..proxies import current_rdm_records_service as service
 from ..services.errors import InvalidAccessRestrictions
@@ -65,7 +72,13 @@ class AcceptAction(actions.AcceptAction):
         # not be immediately visible in the community's records, when the `all versions`
         # facet is not toggled
         uow.register(RecordIndexOp(record, indexer=service.indexer, index_refresh=True))
-
+        uow.register(
+            NotificationOp(
+                CommunityInclusionAcceptNotificationBuilder.build(
+                    identity=identity, request=self.request
+                )
+            )
+        )
         super().execute(identity, uow)