+++ invenio_rdm_records/services/permissions.py
@@ -96,7 +96,6 @@ class RDMRecordPermissionPolicy(RecordPermissionPolicy):
     #
     # Allow searching of records
     can_search = can_all
-    can_search_all = [UserManager, SystemProcess()]
 
     # Allow reading metadata of a record
     can_read = [
@@ -112,6 +111,8 @@ class RDMRecordPermissionPolicy(RecordPermissionPolicy):
             else_=can_read,
         )
     ]
+    can_read_deleted_files = can_read_deleted
+    can_media_read_deleted_files = can_read_deleted_files
     # Allow reading the files of a record
     can_read_files = [
         IfRestricted("files", then_=can_view, else_=can_all),