+++ tests/resources/test_files.py
@@ -23,6 +23,13 @@ def create_draft(client, record, headers):
     return response.json
 
 
+def delete_record(client, recid, headers):
+    """Soft delete a record."""
+    response = client.delete(f"/records/{recid}/delete", json={}, headers=headers)
+    assert response.status_code == 204
+    response.close()
+
+
 def attach_file(client, recid, key, headers):
     """Attach a file to a record."""
 
@@ -108,3 +115,102 @@ def test_published_record_files_deny_edit(
     # the upload of a file is the one that is affected when we lock the bucket
     assert response.status_code == 403
     logout_user(client)
+
+
+def test_files_api_flow_for_deleted_record(
+    client, headers, running_app, minimal_record, users, location, superuser
+):
+    login_user(client, users[0])
+
+    draft = create_draft(client, minimal_record, headers)
+    recid = draft["id"]
+
+    attach_file(client, recid, "test.pdf", headers)
+
+    # publish the draft
+    response = client.post(link(draft["links"]["publish"]), headers=headers)
+    assert response.status_code == 202
+    assert response.data
+
+    logout_user(client)
+
+    url = f"/records/{recid}/files"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files/test.pdf"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files/test.pdf/content"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files-archive"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    login_user(client, superuser.user)
+
+    # We delete the record
+    delete_record(client, recid, headers)
+
+    # Superuser has access to record
+    url = f"/records/{recid}/files"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files/test.pdf"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files/test.pdf/content"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files-archive"
+
+    response = client.get(url, headers=headers)
+    assert 200 == response.status_code
+    assert response.data
+
+    logout_user(client)
+
+    # Non superuser users have no access to the record
+    url = f"/records/{recid}/files"
+
+    response = client.get(url, headers=headers)
+    assert 410 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files/test.pdf"
+
+    response = client.get(url, headers=headers)
+    assert 410 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files/test.pdf/content"
+
+    response = client.get(url, headers=headers)
+    assert 410 == response.status_code
+    assert response.data
+
+    url = f"/records/{recid}/files-archive"
+
+    response = client.get(url, headers=headers)
+    assert 410 == response.status_code
+    assert response.data